changing next-auth version to use the last one that is v5
