+++ package.json
@@ -18,7 +18,7 @@
     "clsx": "^2.1.1",
     "lucide-react": "^0.414.0",
     "next": "14.2.5",
-    "next-auth": "^4.24.7",
+    "next-auth": "^5.0.0-beta.19",
     "next-themes": "^0.3.0",
     "nodemailer": "^6.9.14",
     "react": "^18",
