+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,1 @@
-import { handlers } from "@/server/auth";
-
-export const { GET, POST } = handlers;
+export { GET, POST } from "@/server/auth"
